refactoring main/views
